+++ imageflow/lib/Controller/HealthController.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace OCA\ImageFlow\Controller;
 
 use OCA\ImageFlow\AppInfo\Application;
+use OCA\ImageFlow\Service\BackgroundGateService;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http\Attribute\NoAdminRequired;
 use OCP\AppFramework\Http\JSONResponse;
@@ -15,6 +16,7 @@ class HealthController extends Controller {
 	public function __construct(
 		IRequest $request,
 		private readonly IConfig $config,
+		private readonly BackgroundGateService $backgroundGateService,
 	) {
 		parent::__construct(Application::APP_ID, $request);
 	}
@@ -23,6 +25,7 @@ class HealthController extends Controller {
 	public function index(): JSONResponse {
 		$realExecutionEnabled = $this->config->getAppValue(Application::APP_ID, 'real_execution_enabled', '0') === '1';
 		$backgroundProcessingEnabled = $this->config->getAppValue(Application::APP_ID, 'background_processing_enabled', '0') === '1';
+		$backgroundGate = $this->backgroundGateService->status();
 
 		return new JSONResponse([
 			'app' => Application::APP_ID,
@@ -34,6 +37,7 @@ class HealthController extends Controller {
 			'destructiveWritesEnabled' => $realExecutionEnabled,
 			'realExecutionEnabled' => $realExecutionEnabled,
 			'backgroundProcessingEnabled' => $backgroundProcessingEnabled,
+			'backgroundGate' => $backgroundGate,
 			'safeModeDefault' => true,
 		]);
 	}